upload lab3

+++ lab3/Flik/Flik.java
@@ -1,6 +1,11 @@
-/** An Integer tester created by Flik Enterprises. */
+import static org.junit.Assert.*;
+import org.junit.Test;
+
+/**
+ * An Integer tester created by Flik Enterprises.
+ */
 public class Flik {
     public static boolean isSameNumber(Integer a, Integer b) {
-        return a == b;
+        return (int) a == (int) b;
     }
 }
